add tests for form group
